pyproject.toml
[tool.poetry]
name = "pdbst-rtree"
version = "0.1.0"
description = ""
authors = ["LionTao <taojiachun980831@163.com>"]

[tool.poetry.dependencies]
python = "^3.8"
dapr = "^1.5.0"
dapr-ext-grpc = "^1.5.0"
prettyprint = "^0.1.5"
aiologger = "^0.6.1"
dapr-ext-fastapi = "^1.5.0"
dacite = "^1.6.0"
split = "^0.4"
diagrams = "^0.21.1"
jupyter = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
ipykernel = "^6.9.2"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
